+++ client/components/index.js
@@ -4,12 +4,12 @@
  * having to figure out which file they belong to!
  */
 export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
-export {default as SingleHome} from './single-home'
-export {default as HomeInfo} from './home-info'
-export {default as HomeTitle} from './home-title'
-export {default as HomeSearch} from './Home-Search'
-export {default as HomeSearchCard} from './Home-Search-Card'
-export {default as HomeContent} from './home-content'
+export {default as UserHome} from './user/user-home'
+export {default as SingleHome} from './singleHome/single-home'
+export {default as HomeInfo} from './singleHome/home-info'
+export {default as HomeTitle} from './singleHome/home-title'
+export {default as HomeSearch} from './filterHomes/home-search'
+export {default as HomeSearchCard} from './filterHomes/home-search-card'
+export {default as HomeContent} from './singleHome/home-content'
 export {default as Login} from './login-form'
 export {default as Signup} from './signup-form'
