+++ package.json
@@ -41,6 +41,7 @@
     "connect-session-sequelize": "^4.1.0",
     "express": "^4.14.1",
     "express-session": "^1.15.1",
+    "fs": "0.0.1-security",
     "history": "^4.6.3",
     "morgan": "^1.8.1",
     "passport": "^0.3.2",
@@ -78,13 +79,13 @@
     "husky": "^0.14.3",
     "lint-staged": "^7.0.1",
     "mocha": "^3.3.0",
-    "nodemon": "^1.12.1",
+    "nodemon": "^1.17.5",
     "prettier": "1.11.1",
     "react-test-renderer": "^16.0.0",
     "redux-devtools-extension": "^2.13.2",
     "redux-mock-store": "^1.2.3",
     "supertest": "^3.0.0",
-    "webpack": "^4.0.1",
+    "webpack": "^4.11.1",
     "webpack-cli": "^2.0.9"
   }
 }
